feat: add lingui use js
