+++ locale/en.json
@@ -6,6 +6,5 @@
   "Lottery": "Lottery",
   "Markting": "Markting",
   "SOLUTIONS": "SOLUTIONS",
-  "SUPPORT": "SUPPORT",
-  "test": "test"
+  "SUPPORT": "SUPPORT"
 }
