+++ locale/fr.js
@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:{"Blockchain lottery building service for municipalities.":"Blockchain lottery building service for municipalities.","Dashboard":"Dashboard","LEGAL":"LEGAL","Language":"Language","Lottery":"Lottery","Markting":"Markting","SOLUTIONS":"SOLUTIONS","SUPPORT":"SUPPORT","test":"test"}};
+/*eslint-disable*/module.exports={messages:{"Blockchain lottery building service for municipalities.":"Blockchain lottery building service for municipalities.","Dashboard":"Dashboard","LEGAL":"LEGAL","Language":"Language","Lottery":"Lottery","Markting":"Markting","SOLUTIONS":"SOLUTIONS","SUPPORT":"SUPPORT"}};
