+++ locale/ja.js
@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:{"Blockchain lottery building service for municipalities.":"自治体向けのブロックチェーン宝くじ構築サービス","Dashboard":"ダッシュボード","LEGAL":"法律","Language":"言語","Lottery":"宝くじ","Markting":"マーケティング","SOLUTIONS":"ソリューション","SUPPORT":"サポート","test":"テスト"}};
+/*eslint-disable*/module.exports={messages:{"Blockchain lottery building service for municipalities.":"自治体向けのブロックチェーン宝くじ構築サービス","Dashboard":"ダッシュボード","LEGAL":"法律","Language":"言語","Lottery":"宝くじ","Markting":"マーケティング","SOLUTIONS":"ソリューション","SUPPORT":"サポート"}};
