+++ locale/ja.json
@@ -6,6 +6,5 @@
   "Lottery": "宝くじ",
   "Markting": "マーケティング",
   "SOLUTIONS": "ソリューション",
-  "SUPPORT": "サポート",
-  "test": "テスト"
+  "SUPPORT": "サポート"
 }
