+++ locale/vi.json
@@ -6,6 +6,5 @@
   "Lottery": "",
   "Markting": "",
   "SOLUTIONS": "",
-  "SUPPORT": "",
-  "test": ""
+  "SUPPORT": ""
 }
