+++ src/config/index.ts
@@ -1,87 +1,6 @@
-import { t } from '@lingui/macro'
-
 const config = {
   serviceName: 'Foxlottery',
   serviceIcon: `<svg viewBox="0 0 500 500" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M428.712 82.7482L249.981 499.994L71.3004 82.7906C93.0499 104.902 120.349 122.702 151.244 134.921C182.14 147.14 215.868 153.473 250.018 153.47C284.168 153.465 317.895 147.123 348.787 134.898C379.678 122.672 406.971 104.864 428.712 82.7482Z" fill="url(#paint0_linear_2_77)"/><g style="mix-blend-mode:multiply" opacity="0.08"><g style="mix-blend-mode:multiply" opacity="0.08"><path d="M428.712 82.7482L249.981 499.994L71.3004 82.7906C93.0159 104.927 120.303 122.746 151.198 134.967C182.094 147.188 215.83 153.507 249.981 153.469C185.323 153.469 149.662 146.946 130.217 140.503L249.981 470.433L372.332 139.608C353.543 146.305 317.579 153.469 249.981 153.469C284.144 153.503 317.892 147.18 348.796 134.951C379.7 122.723 406.993 104.894 428.712 82.7482Z" fill="#77C159"/></g></g><path d="M466.899 252.093L249.977 499.995L265.517 463.736C265.574 463.644 265.608 463.543 265.618 463.439L428.708 82.7487L434.191 69.8678L462.735 3.28512L463.952 0.471602L464.161 0L464.261 0.529895L465.025 3.64014L469.29 21.9417L485.997 93.2188C486.006 93.4191 486.039 93.6178 486.098 93.8123L496.824 139.497C498.712 147.508 499.748 155.643 499.922 163.807C500.875 195.497 489.258 226.554 466.899 252.093Z" fill="url(#paint1_linear_2_77)"/><g style="mix-blend-mode:screen" opacity="0.25"><g style="mix-blend-mode:screen" opacity="0.25"><path d="M481.982 120.284L283.187 447.612L438.252 85.6471L443.735 72.7237L461.256 31.9243L481.982 120.284Z" fill="white"/></g></g><g style="mix-blend-mode:screen" opacity="0.3"><g style="mix-blend-mode:screen" opacity="0.3"><path d="M486.097 93.7964C481.933 78.2714 477.895 62.704 473.908 47.0889C470.406 33.6992 466.968 20.3044 463.649 6.86708L442.197 58.8201L417.717 116.914L393.186 174.966C385.06 194.333 376.523 213.572 368.195 232.891L343.121 290.773C334.843 310.097 326.086 329.289 317.625 348.528C308.994 367.762 300.59 387.044 291.826 406.198C283.063 425.353 274.457 444.629 265.516 463.736C265.573 463.644 265.607 463.543 265.617 463.439L428.708 82.7487L434.19 69.8677L462.734 3.28512C462.481 2.30488 462.28 1.32467 462.027 0.344424L463.952 0.471602L464.16 0L464.261 0.529895L466.242 0.657073L465.024 3.64014L469.289 21.9417L485.997 93.2188C486.007 93.4138 486.04 93.6072 486.097 93.7964Z" fill="white"/></g></g><g style="mix-blend-mode:multiply" opacity="0.2"><g style="mix-blend-mode:multiply" opacity="0.2"><path d="M466.899 252.092L249.977 499.994C258.356 489.54 266.937 479.176 275.473 468.812C284.01 458.448 292.691 448.147 301.342 437.889L327.399 407.051L353.709 376.382C371.331 356.036 389.041 335.647 406.884 315.448L433.705 285.124C442.74 275.057 451.681 264.947 460.835 254.922L460.886 254.837L460.936 254.795C464.036 251.636 466.92 248.331 469.574 244.897C472.268 241.4 474.804 237.86 477.145 234.194C479.486 230.527 481.669 226.813 483.701 223.019L486.546 217.259C487.511 215.341 488.269 213.376 489.133 211.431C489.997 209.487 490.811 207.504 491.524 205.502C492.237 203.499 493.045 201.539 493.657 199.531L495.43 193.432C496.047 191.407 496.54 189.358 496.906 187.291C497.31 185.24 497.922 183.194 498.168 181.107L499.032 174.875C499.594 171.209 499.695 167.542 499.947 163.828C500.888 195.513 489.262 226.562 466.899 252.092Z" fill="#77C159"/></g></g><path d="M33.0777 252.093L250.006 500L234.46 463.741C234.403 463.649 234.368 463.548 234.359 463.445L71.2748 82.7487L65.7856 69.8677L37.2419 3.28512L36.0242 0.471602L35.8222 0L35.7213 0.529895L34.9579 3.64014L30.6928 21.9417L13.9794 93.2029C13.9705 93.4032 13.9367 93.6019 13.8784 93.7964L3.15248 139.481C1.26979 147.492 0.235236 155.627 0.0609152 163.791C-0.899237 195.486 10.7155 226.549 33.0777 252.093Z" fill="url(#paint2_linear_2_77)"/><g style="mix-blend-mode:screen" opacity="0.25"><g style="mix-blend-mode:screen" opacity="0.25"><path d="M17.9915 120.284L216.786 447.612L61.7215 85.6524L56.2387 72.729L38.7177 31.9295L17.9915 120.284Z" fill="white"/></g></g><g style="mix-blend-mode:screen" opacity="0.3"><g style="mix-blend-mode:screen" opacity="0.3"><path d="M13.8757 93.7964C18.0398 78.2714 22.0777 62.704 26.0652 47.0889C29.5732 33.701 32.993 20.2938 36.3243 6.86708L57.7761 58.8201L82.2628 116.914L106.794 174.966C114.92 194.333 123.45 213.572 131.779 232.891L156.852 290.773C165.13 310.097 173.887 329.289 182.348 348.528C190.985 367.762 199.383 387.044 208.153 406.198C216.923 425.353 225.523 444.629 234.457 463.736C234.4 463.644 234.365 463.543 234.356 463.439L71.272 82.7487L65.7828 69.8677L37.2391 3.28512C37.4915 2.30488 37.6997 1.32467 37.9521 0.344424L36.0215 0.471602L35.8196 0L35.7186 0.529895L33.7375 0.657073L34.9552 3.64014L30.6901 21.9417L13.9766 93.2029C13.9678 93.4032 13.934 93.6019 13.8757 93.7964Z" fill="white"/></g></g><g style="mix-blend-mode:multiply" opacity="0.2"><g style="mix-blend-mode:multiply" opacity="0.2"><path d="M33.0768 252.093L250.005 500C241.62 489.546 233.039 479.182 224.509 468.818C215.978 458.453 207.291 448.153 198.641 437.895L172.589 407.057L146.279 376.388C128.65 356.042 110.978 335.652 93.0972 315.454L66.2825 285.13C57.2415 275.063 48.3012 264.953 39.1527 254.928L39.1022 254.843L39.0517 254.801C35.9517 251.643 33.0699 248.338 30.4205 244.903C27.7264 241.406 25.1838 237.866 22.8493 234.199C20.5149 230.533 18.3255 226.818 16.2939 223.025L13.4547 217.265C12.4894 215.347 11.7259 213.381 10.8616 211.437C9.99718 209.492 9.18958 207.51 8.47662 205.507C7.76367 203.504 6.94976 201.544 6.34407 199.536L4.56482 193.437C3.94687 191.413 3.45408 189.363 3.08843 187.296C2.68463 185.245 2.07262 183.2 1.82656 181.112L0.962169 174.881C0.406948 171.215 0.331254 167.548 0.0536427 163.834C-0.893428 195.515 10.723 226.561 33.0768 252.093Z" fill="#77C159"/></g></g><defs><linearGradient id="paint0_linear_2_77" x1="-659.964" y1="2434.76" x2="-659.964" y2="2103.89" gradientUnits="userSpaceOnUse"><stop offset="0.160241" stop-color="#77C159"/><stop offset="0.243575" stop-color="#77C159"/><stop offset="0.342535" stop-color="#5FA742"/></linearGradient><linearGradient id="paint1_linear_2_77" x1="-535.27" y1="1949.98" x2="-535.27" y2="2443.47" gradientUnits="userSpaceOnUse"><stop stop-color="#77C159"/><stop offset="0.162" stop-color="#FBC000"/><stop offset="0.307" stop-color="#FBB500"/><stop offset="0.446" stop-color="#FBA300"/><stop offset="0.581" stop-color="#FA8900"/><stop offset="0.712" stop-color="#FA6800"/><stop offset="0.842" stop-color="#F93F00"/><stop offset="0.967" stop-color="#F81000"/><stop offset="1" stop-color="#F80300"/></linearGradient><linearGradient id="paint2_linear_2_77" x1="-12791.5" y1="1950" x2="-12791.5" y2="2443.5" gradientUnits="userSpaceOnUse"><stop stop-color="#77C159"/><stop offset="0.162" stop-color="#FBC000"/><stop offset="0.307" stop-color="#FBB500"/><stop offset="0.446" stop-color="#FBA300"/><stop offset="0.581" stop-color="#FA8900"/><stop offset="0.712" stop-color="#FA6800"/><stop offset="0.842" stop-color="#F93F00"/><stop offset="0.967" stop-color="#F81000"/><stop offset="1" stop-color="#F80300"/></linearGradient></defs></svg>`,
-  organizationName: 'Foxlottery Foundation',
-  header: {
-    menuItems: [],
-  },
-  footer: {
-    description: t`Blockchain lottery building service for municipalities.`,
-    copyright: '© 2022 Foxlottery Foundation. All rights reserved.',
-    iconLinks: [
-      {
-        key: 'facebook',
-        icon: `<svg fill="currentColor" viewBox="0 0 24 24">
-                    <path
-                    fillRule="evenodd"
-                    d="M22 12c0-5.523-4.477-10-10-10S2 6.477 2 12c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V12h2.54V9.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V12h2.773l-.443 2.89h-2.33v6.988C18.343 21.128 22 16.991 22 12z"
-                    clipRule="evenodd"
-                    />
-                </svg>`,
-        link: '',
-      },
-      {
-        key: 'instagram',
-        icon: `<svg fill="currentColor" viewBox="0 0 24 24">
-                <path
-                    fillRule="evenodd"
-                    d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"
-                    clipRule="evenodd"
-                />
-                </svg>`,
-        link: '',
-      },
-      {
-        key: 'twitter',
-        icon: `<svg fill="currentColor" viewBox="0 0 24 24">
-                    <path d="M8.29 20.251c7.547 0 11.675-6.253 11.675-11.675 0-.178 0-.355-.012-.53A8.348 8.348 0 0022 5.92a8.19 8.19 0 01-2.357.646 4.118 4.118 0 001.804-2.27 8.224 8.224 0 01-2.605.996 4.107 4.107 0 00-6.993 3.743 11.65 11.65 0 01-8.457-4.287 4.106 4.106 0 001.27 5.477A4.072 4.072 0 012.8 9.713v.052a4.105 4.105 0 003.292 4.022 4.095 4.095 0 01-1.853.07 4.108 4.108 0 003.834 2.85A8.233 8.233 0 012 18.407a11.616 11.616 0 006.29 1.84" />
-                </svg>`,
-        link: '',
-      },
-      {
-        key: 'github',
-        icon: `<svg fill="currentColor" viewBox="0 0 24 24">
-                    <path
-                    fillRule="evenodd"
-                    d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
-                    clipRule="evenodd"
-                    />
-                </svg>`,
-        link: '',
-      },
-    ],
-    menus: [
-      {
-        title: t`SOLUTIONS`,
-        menuItems: [
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-        ],
-      },
-      {
-        title: t`SUPPORT`,
-        menuItems: [
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-        ],
-      },
-      {
-        title: t`LEGAL`,
-        menuItems: [
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-          { key: '', title: t`Markting`, link: '' },
-        ],
-      },
-    ],
-  },
   lingui: {
     localeKeyValues: {
       en: 'English',
@@ -105,4 +24,5 @@ const config = {
     },
   },
 }
+
 export default config
