+++ src/pages/dashboard/index.tsx
@@ -1,7 +1,10 @@
+import { t } from '@lingui/macro'
+import { useLingui } from '@lingui/react'
 import type { NextPage } from 'next'
 
 const Dashboard: NextPage = () => {
-  return <>dashboard</>
+  const { i18n } = useLingui()
+  return <>{i18n._(t`Dashboard`)}</>
 }
 
 export default Dashboard
