+++ src/pages/index.tsx
@@ -4,12 +4,7 @@ import type { NextPage } from 'next'
 
 const Home: NextPage = () => {
   const { i18n } = useLingui()
-  return (
-    <>
-      Welcome to <a href="https://nextjs.org">Next.js!</a>
-      {i18n._(t`test`)}
-    </>
-  )
+  return <>{i18n._(t`Lottery`)}</>
 }
 
 export default Home
