+++ src/types/IconLink.ts
@@ -1,0 +1,5 @@
+export type IconLink = {
+  key: string
+  link: string
+  icon: string
+}
