+++ tsconfig.json
@@ -7,6 +7,7 @@
       "app/hooks/*": ["./hooks/*"],
       "app/config": ["./config"],
       "app/config/*": ["./config/*"],
+      "app/types/*": ["./types/*"],
     },
     "target": "es2018",
     "lib": ["dom", "dom.iterable", "esnext"],
